Done with the sections of the websit
For now all the basics of each section was done as per Figma. We will be left with adding some images(leaves and flowers), fonts and probably some small animations.

You can edit the code as you see fit to make it better.

If you have questions on the code you can ask me.